3-step
1. Ended Rotes.
2. Change .js to .jsx.
3. Transferred functions from APP to components.
4. Made page Contacts.
