+++ src/components/App.jsx
@@ -1,89 +1,32 @@
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
 import css from './App.module.css';
-// import { ContactForm } from './ContactForm/ContactForm';
-import { Filter } from './Filter/Filter';
-import { ContactList } from './ContactList/ContactList';
-import { Contact } from './Contact/Contact';
-import { fetchContacts, addContact, deleteContact } from 'redux/operations';
-import { chngFilter } from 'redux/filterSlice';
-import {
-  getContacts,
-  getFilter,
-  getIsLoading,
-  getError,
-} from '../redux/selectors';
 import AppBar from './AppBar/AppBar';
 import Register from 'pages/Register/Register';
 import Login from 'pages/Login/Login';
 import AddContact from 'pages/AddContact/AddContact';
+import Contacts from 'pages/Contacts/Contacts';
 
 export function App() {
-  const contacts = useSelector(getContacts);
-  const filter = useSelector(getFilter);
-  const isLoading = useSelector(getIsLoading);
-  const error = useSelector(getError);
-
-  const dispatch = useDispatch();
-
-  const handleSubmit = contact => {
-    if (isSaved(contact)) {
-      return alert(`${contact.name} is already is contacts `);
-    }
-    dispatch(addContact(contact));
-  };
-
-  const isSaved = user => {
-    const normalaseUser = user.name.toLowerCase();
-    return contacts.find(
-      contact => contact.name.toLowerCase() === normalaseUser
-    );
-  };
-
-  const changeFilter = evt => {
-    dispatch(chngFilter(evt.currentTarget.value));
-  };
-
-  const delContact = userId => {
-    dispatch(deleteContact(userId));
-  };
-
-  const filtredContacts = () => {
-    const normalaseFilter = filter.toLowerCase();
-    return contacts.filter(contact =>
-      contact.name.toLowerCase().includes(normalaseFilter)
-    );
-  };
-
-  useEffect(() => {
-    dispatch(fetchContacts());
-  }, [dispatch]);
-
   return (
     <div className={css.mainDiv}>
       <AppBar />
       <Routes>
         <Route path="/register" element={<Register />} />
         <Route path="/login" element={<Login />} />
-        <Route
-          path="/addcontact"
-          element={<AddContact onSubmit={handleSubmit} />}
-        />
+        <Route path="/addcontact" element={<AddContact />} />
+        <Route path="/contacts" element={<Contacts />} />
       </Routes>
       {/* <h2>Phonebook</h2> */}
 
-      {/* <ContactForm onSubmit={handleSubmit} /> */}
+      {/* <h2>Contacts</h2> */}
 
-      <h2>Contacts</h2>
+      {/* <Filter onChange={changeFilter} value={filter} /> */}
 
-      <Filter onChange={changeFilter} value={filter} />
-
-      <ContactList>
+      {/* <ContactList>
         {isLoading && <p>Loading contacts...</p>}
         {error && <p>{error}</p>}
         <Contact contactList={filtredContacts()} onDeleteContact={delContact} />
-      </ContactList>
+      </ContactList> */}
     </div>
   );
 }
