--- src/pages/Contacts/Contacts.js
@@ -1,11 +1,0 @@
-import { ContactForm } from '../../components/ContactForm/ContactForm';
-
-const ContactList = () => {
-  return (
-    <>
-      <ContactForm />
-    </>
-  );
-};
-
-export default ContactList;
